refactor: Restaurant, RestaurantList 관련 리팩터링
