+++ src/data/CategoryImagePath.ts
@@ -4,9 +4,8 @@ import japaneseImage from '../../templates/category-japanese.png';
 import koreanImage from '../../templates/category-korean.png';
 import westernImage from '../../templates/category-western.png';
 import etcImage from '../../templates/category-etc.png';
-import { Category } from './type';
 
-const categoryImages = {
+export const CategoryImagePath = {
   한식: koreanImage,
   중식: chineseImage,
   일식: japaneseImage,
@@ -14,7 +13,3 @@ const categoryImages = {
   양식: westernImage,
   기타: etcImage,
 };
-
-const getCategoryImage = (category: Category) => categoryImages[category];
-
-export { getCategoryImage };
