--- src/data/type.ts
@@ -1,5 +1,0 @@
-type Category = '한식' | '중식' | '일식' | '아시안' | '양식' | '기타';
-type MinutesToCampus = 5 | 10 | 15 | 20 | 30;
-type SortCondition = '이름' | '거리';
-
-export { Category, MinutesToCampus, SortCondition };
